styles: navbar and home page blocks
