+++ src/components/Navbar/ChartLayouts/NavbarChartsDropdown.js
@@ -8,7 +8,31 @@ import { prepareTemplateLink } from '../../../ducks/Studio/Template/utils'
 import { getRecentTemplates } from '../../../utils/recent'
 import { useRecentTemplates } from '../../../hooks/recents'
 import NavbarChartsLayouts from './NavbarChartsLayouts'
+import NewLabel from '../../NewLabel/NewLabel'
 import styles from './NavbarChartsDropdown.module.scss'
+
+const DASHBOARDS = [
+  {
+    name: 'Stablecoins',
+    to: '/stablecoins',
+    createdAt: '2020-10-01T00:00:00Z'
+  },
+  {
+    name: 'Uniswap Protocol',
+    to: '/uniswap-protocol',
+    createdAt: '2020-11-01T00:00:00Z'
+  },
+  {
+    name: 'Decentralized Exchanges',
+    to: '/decentralized-exchanges',
+    createdAt: '2020-11-03T00:00:00Z'
+  },
+  {
+    name: 'Bitcoin Locked on Ethereum',
+    to: '/bitcoin-locked-on-ethereum',
+    createdAt: '2020-11-04T00:00:00Z'
+  }
+]
 
 const NavbarChartsDropdown = ({ activeLink }) => {
   const [layouts = []] = useFeaturedTemplates()
@@ -18,6 +42,37 @@ const NavbarChartsDropdown = ({ activeLink }) => {
   return (
     <Panel>
       <div className={styles.wrapper}>
+        <div className={cx(styles.block, styles.list)}>
+          <h3 className={styles.title}>Dashboards</h3>
+
+          <div className={styles.featuredWrapper}>
+            {DASHBOARDS.map(({ to, name, createdAt }) => {
+              const link = to
+
+              return (
+                <Button
+                  fluid
+                  variant='ghost'
+                  key={name}
+                  as={Link}
+                  to={link}
+                  isActive={link === activeLink}
+                  className={styles.btn}
+                >
+                  {[
+                    <NewLabel
+                      className={styles.new}
+                      date={createdAt}
+                      limitDays={14}
+                      key='new'
+                    />,
+                    name
+                  ]}
+                </Button>
+              )
+            })}
+          </div>
+        </div>
         <div className={styles.block}>
           <h3 className={styles.title}>Explore chart layouts</h3>
           <div className={styles.featuredWrapper}>
